Move ViewportTests to separate file: PageUpDownTests
